Rename "base" service to "threshold_service" in case we have services later that aren't strictly for checking thresholds.
