+++ blottertrax/services/lastfm.py
@@ -6,7 +6,7 @@ from blottertrax.exceptions.empty_description import EmptyDescription
 from blottertrax.exceptions.service_requires_parsed_submission import ServiceRequiresParsedSubmission
 from blottertrax.helper import templates
 from blottertrax.config import Config
-from blottertrax.services.base import ThresholdService
+from blottertrax.services.threshold_service import ThresholdService
 from blottertrax.value_objects.parsed_submission import ParsedSubmission
 from blottertrax.value_objects.service_result import ThresholdServiceResult
