+++ blottertrax/services/soundcloud.py
@@ -1,7 +1,7 @@
 import requests
 
 from blottertrax.config import Config
-from blottertrax.services.base import ThresholdService
+from blottertrax.services.threshold_service import ThresholdService
 from blottertrax.value_objects.parsed_submission import ParsedSubmission
 from blottertrax.value_objects.service_result import ThresholdServiceResult
